requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos[dbt-snowflake]>=1.5.0
apache-airflow-providers-snowflake>=5.0.0
dbt-snowflake>=1.7.0
